Revert "feat: Update Disease Category"
