+++ src/interface/disease/disease-category.interface.ts
@@ -4,6 +4,5 @@ import mongoose from "mongoose";
 export interface IDiseaseCategory {
   name: DiseaseCategoryEnum;
   icon: string;
-  disUsage: mongoose.Types.ObjectId[];
   disease: mongoose.Types.ObjectId[];
 }
