+++ src/middleware/error.middleware.ts
@@ -3,7 +3,7 @@ import { Request, Response, NextFunction } from "express";
 
 const asyncError = (fn: Function) => {
   return (req: Request, res: Response, next: NextFunction) => {
-    Promise.resolve(fn(req, res, next)).catch(next);
+    Promise.resolve(fn(req, res, next).catch(next));
   };
 };
