+++ src/model/disease/disease-usage.model.ts
@@ -1,3 +1,4 @@
+
 import { NameEnum } from "../../enum/disease/disease-usage.enum";
 import { IDiseaseUsageGroup } from "../../interface/disease/disease-usage.interface";
 import mongoose, { Schema } from "mongoose";
@@ -15,20 +16,13 @@ const DiseaseUsageGroupSchema = new Schema<IDiseaseUsageGroup>(
     diseaseCategory: [
       {
         type: Schema.Types.ObjectId,
-        ref: "Disease Category",
-      },
-    ],
-    disease: [
-      {
-        type: Schema.Types.ObjectId,
-        ref: "Disease",
       },
     ],
   },
   { collection: "Disease Usage Group", timestamps: true }
 );
-const DisUsageGroup = mongoose.model(
+const DiseaseUsageGroup = mongoose.model(
   "Disease Usage Group",
   DiseaseUsageGroupSchema
 );
-export default DisUsageGroup;
+export default DiseaseUsageGroup;
