+++ src/router/disease.route.ts
@@ -1,8 +1,5 @@
 import express from "express";
-import controller from "../controller/disease/disease.controller";
+import controller from "../controller/disease.controller";
 const router = express.Router();
 router.post("/add-disease", controller.addDisease);
 router.put("/edit-disease", controller.editDisease);
-router.delete("/delete-disease", controller.editDisease);
-router.get("/get-disease", controller.editDisease);
-export default router;
